fix: contract placeholders, indexer privacy, progress clamp, server cache split
Reject CAAAAA placeholder contract addresses before wallet signing, gate
indexer logs in production, keep step progress monotonic, and split
IndexedDB caching from the server-safe /api/wrapped indexer path.
