+++ .husky/pre-commit
@@ -1,8 +1,16 @@
 #!/bin/sh
 
-
 echo "🔍 Running lint before commit..."
-pnpm lint
+
+# Lint staged files only — full-repo eslint currently fails on unrelated broken files on main.
+STAGED_FILES=$(git diff --cached --name-only --diff-filter=ACM | grep -E '\.(js|jsx|ts|tsx|mjs)$' || true)
+
+if [ -z "$STAGED_FILES" ]; then
+  echo "No staged JS/TS files to lint."
+  exit 0
+fi
+
+echo "$STAGED_FILES" | xargs pnpm exec eslint
 
 status=$?
 
@@ -11,4 +19,3 @@ if [ $status -ne 0 ]; then
   echo "✖ Lint failed. Fix the issues (format/code) before committing."
   exit $status
 fi
-
